feat: :sparkles: Add navigation bar meny and extracted preferences section to another component
